burn-core: Use a specific structure for 1d padding
