+++ burn-core/src/nn/padding.rs
@@ -5,6 +5,35 @@ use burn_tensor::ops::conv::calculate_conv_padding;
 use crate::config::Config;
 use crate::module::Module;
 
+/// Padding configuration for 1D operators.
+#[derive(Module, Config, Debug, PartialEq)]
+pub enum PaddingConfig1d {
+    /// Dynamically calculate the amount of padding necessary to ensure that the output size will be
+    /// the same as the input.
+    Same,
+    /// Same as no padding.
+    Valid,
+    /// Applies the specified amount of padding to all inputs.
+    Explicit(usize),
+}
+
+impl PaddingConfig1d {
+    pub(crate) fn calculate_padding_1d(
+        &self,
+        length: usize,
+        kernel_size: usize,
+        stride: usize,
+    ) -> usize {
+        let same_padding = || calculate_conv_padding(kernel_size, stride, length, length);
+
+        match self {
+            Self::Valid => 0,
+            Self::Same => same_padding(),
+            Self::Explicit(value) => *value,
+        }
+    }
+}
+
 /// Padding configuration for 2D operators.
 #[derive(Module, Config, Debug)]
 pub enum PaddingConfig2d {
